+++ src/UglyToad.PdfPig.Tests/TestObjectLocationProvider.cs
@@ -6,14 +6,14 @@
 
     internal class TestObjectLocationProvider : IObjectLocationProvider
     {
-        public Dictionary<IndirectReference, long> Offsets { get; } = new Dictionary<IndirectReference, long>();
+        public Dictionary<IndirectReference, int> Offsets { get; } = new Dictionary<IndirectReference, int>();
 
-        public bool TryGetOffset(IndirectReference reference, out long offset)
+        public bool TryGetOffset(IndirectReference reference, out int offset)
         {
             return Offsets.TryGetValue(reference, out offset);
         }
 
-        public void UpdateOffset(IndirectReference reference, long offset)
+        public void UpdateOffset(IndirectReference reference, int offset)
         {
             Offsets[reference] = offset;
         }
